add metadata update api request
