+++ .env.production
@@ -39,3 +39,6 @@ REACT_APP_DOMAIN_REGEX_ENS="^cowswap\.eth"
 #REACT_APP_EXPLORER_URL_DEV=https://protocol-explorer.dev.gnosisdev.com
 #REACT_APP_EXPLORER_URL_STAGING=https://protocol-explorer.staging.gnosisdev.com
 #REACT_APP_EXPLORER_URL_PROD=https://gnosis-protocol.io
+
+# Enables mock mode (default = false)
+MOCK=false
