+++ src/custom/utils/operator/operatorMock.ts
@@ -1,0 +1,6 @@
+import { UploadMetadataParams } from './operatorApi'
+
+export async function uploadAppDataDoc(params: UploadMetadataParams): Promise<void> {
+  console.log('[utils:operatorMock] Post AppData doc', params)
+  return
+}
